refactor: shared the sampler's buffer guards with the codebase.
The content lock is now the existing `BorrowedBuffer`. The pin moves to
`heap.rs` as `HeapBufferPin`, kept separate from `PinnedBuffer` so a heap
block never feeds the index `block_tracker`.
